Removed CellBase
